backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
alembic==1.16.4
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
redis==6.2.0
celery==5.5.3
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
qsharp
azure-quantum
azure-identity
azure-keyvault-secrets
opentelemetry-api
opentelemetry-sdk
azure-monitor-opentelemetry
